dom/tests/mochitest/bugs/test_bug918719.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=918719
-->
<head>
  <title>Test for Bug 918719</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=918719">Mozilla Bug 918719</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

const SERVER_URL = "bug918719.sjs";

function sendCommand(cmd) {
  let xhr = new XMLHttpRequest();
  xhr.open("get", SERVER_URL + "?" + cmd);
  xhr.send();
}

function runTest() {
  // Manipulate one in-flight XHR using secondary command XHRs, to guarantee
  // that multiple OnDataAvailable events are triggered (which are where
  // LOADING readystatechanges are triggered). We return a promise that will
  // resolve with a count of the number of LOADING events that were detected.

  return new Promise((resolve, reject) => {
    let xhr = new XMLHttpRequest();
    let numProgressEvents = 0;
    let numLoadingEvents = 0;

    xhr.onreadystatechange = e => {
      if (xhr.readyState === xhr.LOADING) {
        ++numLoadingEvents;
      }
    };

    xhr.onprogress = e => {
      if (++numProgressEvents < 2) {
        sendCommand("more");
      } else {
        sendCommand("done");
      }
    };

    xhr.onerror = e => {
      reject(e);
    };

    xhr.onloadend = e => {
      resolve(numLoadingEvents);
    };

    xhr.open("GET", SERVER_URL);
    xhr.send();
  });
}

function prefChangePromise(args) {
  return new Promise(function(resolve) {
    SpecialPowers.pushPrefEnv(args, resolve);
  });
}

runTest().then(function(count) {
  ok(count === 1, "Only one loading readystatechange event should have been fired with the pref off.");
}).then(function() {
  return prefChangePromise({"set": [["dom.fire_extra_xhr_loading_readystatechanges", true]]});
}).then(function() {
  return runTest();
}).then(function(count) {
  ok(count > 1, "Multiple loading readystatechange events should have been fired with the pref on.");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
